// src/lib/config.d.ts
export declare const appConfig: {
    apiUrl: any;
    web3AuthClientId: any;
    endpoints: {
        payments: string;
        analytics: string;
        users: string;
        templates: string;
    };
};
export declare const getApiHeaders: (userId?: string) => {
    'x-user-id': string;
    'Content-Type': string;
    'ngrok-skip-browser-warning': string;
};
//# sourceMappingURL=config.d.ts.map